Add sinon-qunit package
Sinon provides a well rounded assertion API that by default throws
exceptions. Sinon-qunit package is a plugin-ish to use sinon assertions
but be plugged into QUnit test framework. With the two together you can
get better and more descriptive assertions while working with QUnit.

Since Ember-CLI uses QUnit it would be fantastic if the sinon-qunit
package was available.

Because sinon-qunit requires QUnit to be loaded first (I hate globals)
the only way to make that happen is to lazily load the module. Since the
original sinon-qunit code was never intended for that it was better to
use the AMD shim to manage it. The source was so small it seems an
acceptable option to manually place the code into the shim.

1.  The code is pretty stable and is unlikely to have much churn.
2.  It is wrapped in a conditional so it won't be executed if the user
    chooses not to use QUnit in favor of another test framework.